ICD-10CM codes are now date-ranged (#3)
I am not sure how this will be used on the public website, but I kept
the logic very simple in this project: if a given ICD-10CM code is
valid for any part of a given year, then it's valid for the full year.
It then follows the same logic as any other year-range field.